+++ src/debug.mock.ts
@@ -5,8 +5,7 @@ export type DebugMock = jest.Mocked<Debugger>
 export function mockDebugger(): DebugMock {
     const debug = jest.fn()
 
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
+    // @ts-expect-error
     debug.extend = jest.fn().mockReturnValue(debug)
 
     return debug as Partial<DebugMock> as DebugMock